feat: :sparkles: Add color function and mouseover event listener to grid squares
In this commit, a color function was added to the Etch-a-Sketch application. This function retrieves the current color value from the color picker in the UI. Additionally, a 'mouseover' event listener was added to each grid square (cell) in the application. When the mouse cursor hovers over a grid square, the square's background color changes to the currently selected color in the color picker. This enhancement allows users to 'draw' on the grid by simply moving their mouse over the squares, creating a more interactive and engaging user experience.